Added checks for suspicious activities from logged-in user in users/ app.

+++ users/views.py
@@ -1,4 +1,4 @@
-from django.core.exceptions import PermissionDenied
+from django.core.exceptions import PermissionDenied, SuspiciousOperation
 from django.http import HttpResponseRedirect
 from django.shortcuts import render
 
@@ -10,6 +10,7 @@ from django.views.generic import TemplateView
 from friends.models import Friend
 from main_app.utils import get_friends, get_sent_requests, get_received_requests, get_not_friends
 from users.forms import CustomUserCreationForm
+from users.models import CustomUser
 
 
 class SignUp(generic.CreateView):
@@ -58,6 +59,8 @@ def add_friend(request):
     if not request.user.is_authenticated:
         raise PermissionDenied
     friend_id = request.POST.get('friend', 'default')
+    if not CustomUser.objects.filter(id=friend_id).exists():
+        raise SuspiciousOperation("Please be in limits.")
     Friend.objects.create(creator_id=request.user.id, follower_id=friend_id, confirmed=False)
     return HttpResponseRedirect(reverse('friends:friends'))
 
@@ -67,6 +70,8 @@ def decline(request):
         raise PermissionDenied
     friend_id = request.POST.get('friend', 'default')
     # print(friend)
+    if not CustomUser.objects.filter(id=friend_id).exists():
+        raise SuspiciousOperation("Please be in limits.")
     Friend.objects.filter(creator_id=friend_id, follower_id=request.user.id, confirmed=False).delete()
     return HttpResponseRedirect(reverse('friends:friends'))
 
@@ -76,6 +81,8 @@ def accept(request):
         raise PermissionDenied
     # print("hiiiiiiii")
     friend_id = request.POST.get('friend', 'default')
+    if not CustomUser.objects.filter(id=friend_id).exists():
+        raise SuspiciousOperation("Please be in limits.")
     row = Friend.objects.get(creator_id=friend_id, follower_id=request.user.id, confirmed=False)
     row.confirmed = True
     row.save()
@@ -88,6 +95,8 @@ def cancel(request):
     if not request.user.is_authenticated:
         raise PermissionDenied
     friend_id = request.POST.get('friend', 'default')
+    if not CustomUser.objects.filter(id=friend_id).exists():
+        raise SuspiciousOperation("Please be in limits.")
     Friend.objects.filter(creator_id=request.user.id, follower_id=friend_id).delete()
     return HttpResponseRedirect(reverse('friends:friends'))
 
@@ -96,6 +105,8 @@ def remove_friend(request):
     if not request.user.is_authenticated:
         raise PermissionDenied
     friend_id = request.POST.get('friend', 'default')
+    if not CustomUser.objects.filter(id=friend_id).exists():
+        raise SuspiciousOperation("Please be in limits.")
     row = Friend.objects.filter(creator_id=request.user.id, follower_id=friend_id)
     # print("ro")
     # print(row)
